Powershell script to append .txt file w/Spec info
The various commands for gathering basic consumer specifications of the
current machine. This info is then appended to a text file (hardcoded as
welp.txt) on the current users Desktop. For future reference: Xreate a
selenium script that posts to CraigsList, Amazon, and/or Ebay using the
generated text to fill in the descriptions, where necessary. Otherwise
simply use Make/Model to find the item on Amazon and add to their
preconfigured product catalog.