Rounds total RAM/HDD bytes UP to nearest even #
Rounds the spec info that is gathered in Bytes (RAM and HDD) and rounds
this number UP to the nearest even number. Should be mostly universal
for RAM seeing as how RAM is nearly always an even total number and only
ever displayed (internally) as less. ie commercial info is always listed
as having more than what the computer sees. An issue that still occurs
is when a desktop RAM has been upgraded (desktop) to an odd number. This
method falsely will round up 1 extra GB. Also the same applies for HDD's
over 1000GB. This code is hard coded for testing on a 1TB computer.
Which means that the decimal point will be in the wrong location is ran
on a laptop/comp with < 1000GB. Need to add an `if/else` statement to
change decimal point if over/under X amount of bytes