Relatively quiet (no console) W/ EXE in/output
Used Pyinstaller on main, static, Python script. First file to be
executed is the EXE. the .Py script is exactly the same and used for
verification, authenticity, source code, and/or as an alternative.

Main .exe uses Easy gui to prompt for end user input. This enduser input
is used to create a selenium webdriver script (.py format). The main EXE
then uses PYinstaller and various flags to convert this selenium script
into an EXE. Lastly the main script executes the generated EXE.

All of this is coded for the specific file structure in order to
implement portable python as well as PyInstaller.